flask-image/requirements.txt
alembic>=1.7.4
aniso8601>=8.0.0
bcrypt>=3.2.0
blinker>=1.5
cffi>=1.15.0
click>=8.1.8
cryptography>=35.0.0
dominate>=2.6.0
Flask>=3.1.0
Flask-Bootstrap>=3.3.7.1
Flask-Cors>=1.10.3
Flask-JWT-Extended>=4.1.0
Flask-Login>=0.6.3
Flask-Mail>=0.10.0
Flask-Migrate>=4.1.0
Flask-RESTful>=0.3.10
Flask-Script>=2.0.6
Flask-SQLAlchemy>=2.5.1
Flask-WTF>=1.2.2
flask-admin>=1.6.1
greenlet>=1.1.2
importlib-metadata>=4.8.1
importlib-resources>=5.3.0
itsdangerous>=2.2
Jinja2>=3.1.5
Mako>=1.1.5
Markdown>=3.3.4
MarkupSafe>=1.1.1
paramiko>=2.8.0
psycopg2-binary>=2.8.6
pycparser>=2.20
Pygments>=2.8.1
PyJWT>=2.3.0
PyNaCl>=1.4.0
pytz>=2020.4
PyYAML>=5.4.1
six>=1.15.0
SQLAlchemy>=1.4.26
uWSGI>=2.0.17.1
visitor>=0.1.3
Werkzeug>=3.1.3
WTForms>=2.3.3
zipp>=3.6.0
python-dotenv>=1.0.1
pillow>=10.4.0
gunicorn>=21.2.0
flask-sitemapper>=1.8.1
python-slugify>=8.0.4